requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
gradio>=6.3.0
pypdf>=3.17.0
openai>=1.12.0
